requirements.txt
numpy==1.18.2
paho-mqtt==1.5.0
psycopg2==2.8.5
